fix: performance
