+++ app/blog/[slug]/page.tsx
@@ -1,8 +1,7 @@
 import { notFound } from "next/navigation";
 import { BlogLayout } from "@/components/blog/BlogLayout";
 import BlogPostClient from "@/components/blog/BlogPostClient";
-import type { BlogPost } from "@/types/blog";
-import { getPostBySlug, getAllPosts } from "@/lib/blog";
+import { getAllPosts } from "@/lib/blog";
 
 interface BlogPostPageProps {
   params: Promise<{ slug: string }>;
@@ -17,9 +16,9 @@ export async function generateStaticParams() {
 
 export default async function BlogPost({ params }: BlogPostPageProps) {
   const { slug } = await params;
-  const post = await getPostBySlug(slug);
-  const author = null; // Mock author data
   const allPosts = await getAllPosts();
+  const post = allPosts.find((entry) => entry.slug === slug);
+  const author = null; // Mock author data
 
   if (!post || post.draft) {
     notFound();
